Реализует e2e тест для компонента Navigation
