+++ src/components/header/header.test.js
@@ -1,23 +1,32 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import {Header} from "./header.jsx";
+import configureStore from "redux-mock-store";
+import Header from "./header.jsx";
+import NameSpace from "../../reducer/name-space.js";
+import {AuthorizationStatus} from "../../const.js";
+import {Provider} from 'react-redux';
 
-const AuthInfo = {
-  avatarUrl: `img/1.png`,
-  email: `Oliver.conner@gmail.com`,
-  id: 1,
-  isPro: false,
-  name: `Oliver.conner`
-};
+const mockStore = configureStore([]);
 
 it(`Render Header`, () => {
+  const store = mockStore({
+    [NameSpace.USER]: {
+      authorizationStatus: AuthorizationStatus.NO_AUTH,
+      authInfo: {
+        avatarUrl: ``,
+        name: ``,
+        id: null,
+        isPro: null,
+        email: ``,
+      }
+    }
+  });
+
   const tree = renderer
   .create(
-      <Header
-        isLogoActive={false}
-        authStatus={`AUTH`}
-        authInfo={AuthInfo}
-      />
+      <Provider store={store}>
+        <Header isLogoActive={false} />
+      </Provider>
   ).toJSON();
   expect(tree).toMatchSnapshot();
 });
